Move HTTP request (fetch) logic to separate module
Just to make it clearer to understand what SeedCli is doing, and to remove some of the duplication, we move the logic for sending http requests to the service to a separate module, SystemRequest.
